add router-flux and skeleton for Home and Login page
