+++ package.json
@@ -8,7 +8,12 @@
 	},
 	"dependencies": {
 		"react": "16.0.0",
-		"react-native": "0.50.3"
+		"react-native": "0.50.3",
+		"react-native-router-flux": "^4.0.0-beta.24",
+		"react-redux": "^5.0.6",
+		"redux": "^3.7.2",
+		"redux-logger": "^3.0.6",
+		"redux-thunk": "^2.2.0"
 	},
 	"devDependencies": {
 		"babel-jest": "21.2.0",
